+++ src/products/model.rs
@@ -8,14 +8,14 @@ pub struct Product {
     pub tenant_id: String,
     pub name: String,
     pub sku: String,
-    pub price: f64,
+    pub price: i64,
     /// Purchase price (cost of goods) — the basis for profit report
     /// calculations. Not the selling `price` customers see. Unlike the
     /// rest of this struct, `cost_price` is NOT exposed to every role in
     /// responses: only Owner/Admin get to see it (see `ProductResponse`),
     /// same tier of access as the `/tenants/me/reports/profit` endpoint,
     /// so a Cashier can view a product's catalog data but never its margin.
-    pub cost_price: f64,
+    pub cost_price: i64,
     pub stock: i32,
     pub created_by: Actor,
 }
@@ -32,9 +32,9 @@ pub struct ProductResponse {
     pub tenant_id: String,
     pub name: String,
     pub sku: String,
-    pub price: f64,
+    pub price: i64,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub cost_price: Option<f64>,
+    pub cost_price: Option<i64>,
     pub stock: i32,
     pub created_by: Actor,
 }
@@ -58,8 +58,8 @@ impl ProductResponse {
 pub struct CreateProductRequest {
     pub name: String,
     pub sku: String,
-    pub price: f64,
-    pub cost_price: f64,
+    pub price: i64,
+    pub cost_price: i64,
     pub stock: i32,
 }
 
@@ -70,7 +70,7 @@ pub struct CreateProductRequest {
 #[derive(Debug, Deserialize)]
 pub struct UpdateProductRequest {
     pub name: Option<String>,
-    pub price: Option<f64>,
-    pub cost_price: Option<f64>,
+    pub price: Option<i64>,
+    pub cost_price: Option<i64>,
     pub stock: Option<i32>,
 }
